// src/app/model/interfaces/NomineeDetails.model.ts
export interface NomineeDetails {
    ownerId: number
    nomineeRefNum: number
    ownerType: number
    seqNum: number
    nomShare: number
    nomType: number
    nomTypeCode: number
    nomineeFirstName: string
    nomineeMiddleName: string
    nomineeLastName: string
    rtlnType: number
    rtlnTypeCode: number
    dateOfBirth: string
    nomAddId: number
    nomDocId: string
    ver: number
  }